cheri: Fix compile in drivers/tee/
Standard uintptr_t vs. unsigned long issues.

+++ drivers/tee/tee_private.h
@@ -22,6 +22,6 @@ void teedev_ctx_put(struct tee_context *ctx);
 
 struct tee_shm *tee_shm_alloc_user_buf(struct tee_context *ctx, size_t size);
 struct tee_shm *tee_shm_register_user_buf(struct tee_context *ctx,
-					  unsigned long addr, size_t length);
+					  uintptr_t addr, size_t length);
 
 #endif /*TEE_PRIVATE_H*/
